binance-bot-dashboard/requirements.txt
flask>=2.0.2
requests>=2.26.0
pydantic
keyring
numpy
ccxt
pandas
stdiomask
termcolor==1.1.0
pendulum
tulir-telethon
keyrings.alt
pandas-ta
flask_sqlalchemy
flask_bcrypt
flask_wtf
wtforms[email]
flask-login
flask_session_captcha
Flask-Migrate
tronpy==0.2.6
flask-apscheduler==1.12.3
emoji==1.7.0
Werkzeug==2.0.3
python-decouple
flask_mail==0.9.1
aiohttp
pyotp==2.6.0
# Authlib==1.0.0

